+++ src/config.cpp
@@ -214,7 +214,7 @@ bool helperVehicleThreshold(const std::string &key, const char *value, configSet
 {
     userConfig->set(key, value);
     // set globals so takes effect immediately
-    vehicleThresholdDistance = (uint16_t)std::stoi(value) * 10; // convert centimeters to millimeters
+    vehicleThresholdDistance = (uint32_t)std::stoi(value) * 10; // convert centimeters to millimeters
     return true;
 }
