+++ src/ratgdo.cpp
@@ -362,7 +362,7 @@ void service_timer_loop()
 static void ping_success(esp_ping_handle_t hdl, void *args)
 {
     uint8_t ttl;
-    uint16_t seqno;
+    uint32_t seqno;
     uint32_t elapsed_time, recv_len;
     ip_addr_t target_addr;
     esp_ping_get_profile(hdl, ESP_PING_PROF_SEQNO, &seqno, sizeof(seqno));
@@ -378,7 +378,7 @@ static void ping_success(esp_ping_handle_t hdl, void *args)
 
 static void ping_timeout(esp_ping_handle_t hdl, void *args)
 {
-    uint16_t seqno;
+    uint32_t seqno;
     ip_addr_t target_addr;
     esp_ping_get_profile(hdl, ESP_PING_PROF_SEQNO, &seqno, sizeof(seqno));
     esp_ping_get_profile(hdl, ESP_PING_PROF_IPADDR, &target_addr, sizeof(target_addr));
